+++ src/creatartis-build.js
@@ -1,101 +1,11 @@
 #!/usr/bin/env node
-const { spawn } = require('child_process');
-const { promises: fs } = require('fs');
-const path = require('path');
-
-const packageJSON = require(`${process.cwd()}/package.json`);
-
-// Utilities ///////////////////////////////////////////////////////////////////
-
-async function run(script) {
-  const commands = script.trim().split(/\s*&&\s*|[\r\n]+/)
-    .map((line) => line.trim().split(/[ \t]+/));
-  let result = 0;
-  for (const [command, ...args] of commands) {
-    result = await new Promise((resolve, reject) => {
-      const childProcess = spawn(command, args, {
-        shell: process.platform === 'win32',
-        stdio: 'inherit',
-      });
-      childProcess.on('exit', (code) => resolve(code || childProcess.exitCode));
-      childProcess.on('error', (error) => reject(error));
-    });
-    if (result !== 0) break;
-  }
-  return result;
-}
-
-function distPackageJSON() {
-  const {
-    script, devDependencies, files, ...other
-  } = packageJSON;
-  return JSON.stringify(other, null, '  ');
-}
-
-// Tasks ///////////////////////////////////////////////////////////////////////
-
-function taskPwd() {
-  console.log(`creatartis-build is running at ${process.cwd()}.`);
-}
-
-async function taskLint() {
-  const ignored = ['/dist', '/node_modules', '/docs']
-    .map((pattern) => `--ignore-pattern ${pattern}`)
-    .join(' ');
-  return run(`npx eslint . --ext js,jsx --quiet ${ignored}`);
-}
-
-async function taskTest() {
-  return run(`
-    npx jest ./test/specs --config ${path.join(__dirname, 'jest-config.js')}
-  `);
-}
-
-async function taskBuild(type) {
-  await fs.rmdir('./dist', { recursive: true });
-  await fs.mkdir('./dist');
-  await fs.copyFile('./README.md', './dist/README.md');
-  await fs.copyFile('./LICENSE.md', './dist/LICENSE.md');
-  await fs.writeFile('./dist/package.json', distPackageJSON());
-  let result = 0;
-  if (!type || type === 'umd') {
-    result = await run(`
-      webpack ./src/index.js --config ${path.join(__dirname, 'webpack-config.js')}
-    `);
-    if (result !== 0) return result;
-  }
-  if (!type || type === 'esm') {
-    result = await run('npx babel src/ --out-dir dist/ --source-maps=true');
-  }
-  return result;
-}
-
-async function taskDoc() {
-  await fs.rmdir('./docs/jsdoc', { recursive: true });
-  return run(`
-    npx jsdoc README.md src/ -c ${path.join(__dirname, 'jsdoc-config.js')}
-  `);
-}
-
-function getNPMRegistry(id) {
-  if (!id) return null;
-  if (/^(verdaccio|local)$/.test(id)) return 'http://localhost:4873';
-  if (/^(npm)$/.test(id)) return 'https://registry.npmjs.org/';
-  if (/^\d+/.test(id)) return `http://localhost:${id}`;
-  throw new Error(`Unrecognized NPM registry ${id}!`);
-}
-
-async function taskRelease(id) {
-  const registryURL = getNPMRegistry(id);
-  const registry = registryURL ? `--registry ${registryURL}` : '';
-  return run(`npm publish ${registry} ./dist`);
-}
-
-async function taskUnrelease(id) {
-  const registryURL = getNPMRegistry(id);
-  const registry = registryURL ? `--registry ${registryURL}` : '';
-  return run(`npm unpublish ${registry} --force`);
-}
+const { run, log } = require('./tasks/common');
+const { taskBuild } = require('./tasks/builds');
+const { taskLint } = require('./tasks/checks');
+const { taskDoc } = require('./tasks/docs');
+const { taskPwd } = require('./tasks/other');
+const { taskRelease, taskUnrelease } = require('./tasks/releases');
+const { taskTest } = require('./tasks/tests');
 
 const TASKS = [
   [/pwd/, taskPwd],
@@ -115,7 +25,7 @@ async function execTask(id) {
   for (const [selector, ...taskFunctions] of TASKS) {
     const match = (new RegExp(`^${selector.source}$`)).exec(id);
     if (match) {
-      console.log(`🚧 Executing ${id}...`);
+      log.info(`Executing ${id}...`);
       for (const taskFunction of taskFunctions) {
         result = await taskFunction(...match.slice(1));
         if (result !== 0) break;
@@ -123,7 +33,7 @@ async function execTask(id) {
       return result;
     }
   }
-  console.error(`💥 \x1B[1;31mUnrecognized task ${id}!\x1B[0m`);
+  log.error(`Unrecognized task ${id}!`);
   return 1;
 }
 
@@ -134,7 +44,7 @@ async function main() {
   for (const arg of process.argv.slice(2)) {
     result = await execTask(arg);
     if (result !== 0) {
-      console.error(`💥 \x1B[1;31mTask ${arg} failed! Aborting.\x1B[0m`);
+      log.error(`Task ${arg} failed! Aborting.`);
       break;
     }
   }
