+++ src/tasks/other.js
@@ -1,0 +1,7 @@
+function taskPwd() {
+  console.log(`creatartis-build is running at ${process.cwd()}.`);
+}
+
+module.exports = {
+  taskPwd,
+};
